+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- in `#save` (without bang), `ActiveRecord::RecordInvalid` error is not passed outward.
+
 ## [0.2.0] - 2024-01-21
 
 - add i18n doc.